Change syntax to has_public_id
