+++ lib/public_id/util.rb
@@ -6,7 +6,7 @@ module PublicId
     def self.generate_prefix(klass, joiner)
       klass.to_s.demodulize.underscore.first(3) + joiner
     end
-    def self.new_public_identifier(klass, options = {})
+    def self.new_public_id(klass, options = {})
       length = options[:length] || 10
       prefix = options.fetch(:prefix, generate_prefix(klass, options.fetch(:joiner, '-') ))
       suffix = generate_random_suffix(length)
