+++ lib/public_id/version.rb
@@ -1,3 +1,3 @@
 module PublicId
-  VERSION = "0.5.0"
+  VERSION = "1.0.0"
 end
